requirements.txt
aiogram==2.21
aiohttp==3.8.1
pytz==2022.1
python-dotenv==0.20.0
